<rdar://problem/13246939>
Make format uint64_t[] actually work as designed

llvm-svn: 178072

+++ lldb/source/Core/DataExtractor.cpp
@@ -1811,7 +1811,7 @@ DataExtractor::Dump (Stream *s,
 
         case eFormatVectorOfUInt64:
             s->PutChar('{');
-            offset = Dump (s, offset, eFormatHex,     sizeof(uint32_t), item_byte_size / sizeof(uint32_t), item_byte_size / sizeof(uint32_t), LLDB_INVALID_ADDRESS, 0, 0);
+            offset = Dump (s, offset, eFormatHex,     sizeof(uint64_t), item_byte_size / sizeof(uint64_t), item_byte_size / sizeof(uint64_t), LLDB_INVALID_ADDRESS, 0, 0);
             s->PutChar('}');
             break;
